Modify SELECT and TIME tests
Although `SELECT` with missing `FROM` is not support by
SQL-92 but it's still supported by major SQL DBMS like
PostgresSQL and MySQL. Considering this, moving these
tests as under PartiQL namespace as success tests.

We need to decide further on how we need to proceed forward
with this in our Kotlin reference implementation since it's
currently disallowed.

This PR also fixes some of TIME related tests.